+++ app/src/main/java/edu/ucne/loginapi/domain/model/UserCar.kt
@@ -9,7 +9,7 @@ data class UserCar(
     val fuelType: FuelType,
     val usageType: UsageType,
     val isCurrent: Boolean,
-    val remoteId: Long? = null
+    val remoteId: Int? = null
 )
 
 enum class FuelType {
